pyproject.toml
[project]
name = "agnesagent"
version = "0.0.1"
description = "Highly extensible, cross-platform AI agent infrastructure"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pyyaml>=6.0.1",
    "openai>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "nicegui>=1.4.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
openvino = [
    "openvino>=2024.0.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
]
whisper = [
    "whisper>=1.1.10",
]
dev = [
    "ruff>=0.2.0",
    "pre-commit>=3.0.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "tencent"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.hatch.build.targets.wheel]
packages = ["agnes"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501", "W293", "W291", "E402", "E722", "F401", "F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
typeCheckingMode = "basic"
pythonPlatform = "All"
pythonVersion = "3.12"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportMissingImports = false
reportAttributeAccessIssue = false
reportUnboundVariable = false
reportCallIssue = false
reportArgumentType = false
reportOptionalMemberAccess = false
reportOptionalCall = false

[dependency-groups]
dev = [
    "basedpyright>=1.38.3",
]